+++ axum/src/routing/tests/handle_error.rs
@@ -95,3 +95,17 @@ async fn handler_multiple_methods_last() {
     let res = client.get("/").send().await;
     assert_eq!(res.status(), StatusCode::REQUEST_TIMEOUT);
 }
+
+#[crate::test]
+async fn handler_service_ext() {
+    let fallible_service = tower::service_fn(|_| async { Err::<(), ()>(()) });
+    let handle_error_service =
+        fallible_service.handle_error(|_| async { StatusCode::INTERNAL_SERVER_ERROR });
+
+    let app = Router::new().route("/", get_service(handle_error_service));
+
+    let client = TestClient::new(app);
+
+    let res = client.get("/").send().await;
+    assert_eq!(res.status(), StatusCode::INTERNAL_SERVER_ERROR);
+}